DES function implementation.
DES Encryption has many stages. One of the stages is a function F that takes as its input a block of 32-bits(half of the input block to the DES) and a key that is 48-bits then produces a 32-bit output. The functionality of F can be seen in the following figure: https://codeforces.com/predownloaded/5a/89/5a8926175d76cd7082e4978f89ce3c90f741cd3c.png
The expansion P-box table is given: https://codeforces.com/predownloaded/2c/1e/2c1e739f000d99d89316d9bb788ba6b7cbf05af6.png
The Straight P-box table is given: https://codeforces.com/predownloaded/e8/87/e88740dab06b60c4b92566a8beedbd41efec8417.png
The 32-bit input first pass through an expansion box(Which is just similar to the permutation box you implemented(hopefully) in the first problem.) The output of the expansion box is XORed with the 48-bit key and the result is passed through the 8 S-boxes you implemented in the previous problem to produce 32-bit output that passes through a straight permutation box.

Can you use the codes in your previous submissions to form the full F-Function? Yes, you can.

Input
The input consists of 2 lines. The first line consists of 8 Hex characters that represent the 32-bit input block. The second line consists of 12 Hex characters that represent the 48-bit key block that is also an input to the function.

Output
The output should consist of a single line that has 8 Hex characters which represent the output of the F function given the inputs above.